requirements.txt
streamlit
pandas
psycopg2-binary
sqlalchemy
plotly
numpy
reportlab
